FIR/UAST: commonize expression conversion
