+++ plugins/kotlin/uast/uast-kotlin-base/src/org/jetbrains/uast/kotlin/expressions/KotlinUBinaryExpression.kt
@@ -15,7 +15,7 @@ class KotlinUBinaryExpression(
     givenParent: UElement?
 ) : KotlinAbstractUExpression(givenParent), UBinaryExpression, KotlinUElementWithType, KotlinEvaluatableUElement {
 
-    companion object {
+    internal companion object {
         val BITWISE_OPERATORS = mapOf(
             "or" to UastBinaryOperator.BITWISE_OR,
             "and" to UastBinaryOperator.BITWISE_AND,
